extract generic scroll handler and export from package
