+++ packages/dataviews/src/index.ts
@@ -3,5 +3,6 @@ export { default as DataViewsPicker } from './components/dataviews-picker';
 export { default as DataForm } from './components/dataform';
 export { default as filterSortAndPaginate } from './utils/filter-sort-and-paginate';
 export { useFormValidity } from './hooks';
+export { useInfiniteScrollData } from './dataviews-layouts/utils/useInfiniteScrollData';
 export { VIEW_LAYOUTS } from './dataviews-layouts';
 export type * from './types';
